requirements.txt
matplotlib==3.4.3
numpy==1.21.2
geopandas==0.8.1
rasterio==1.1.8
networkx==2.5
selenium==3.141
cartopy==0.18.0
shapely==1.7.1
rtree==0.9.4
matplotlib-scalebar==0.4.1
pyproj==6.2.1
pyqt==5.9.2
